adding categorie page

+++ stockmanager/src/main/java/com/stock/stockmanager/model/Achat.java
@@ -5,12 +5,9 @@ import java.util.List;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-// import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-// import jakarta.persistence.JoinColumn;
-// import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 
@@ -21,8 +18,7 @@ public class Achat {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_achat")
-    private Long id;
-
+  private Integer id; 
     @Column(name = "date_achat")
     private LocalDate dateAchat;
 
@@ -72,11 +68,10 @@ public class Achat {
         this.receptions = receptions;
     }
 
-    public Long getId() {
+  public Integer getId() {   // <-- getter en Integer
         return id;
     }
-
-    public void setId(Long id) {
+      public void setId(Integer id) {  // <-- setter en Integer
         this.id = id;
     }
 }
