+++ stockmanager/src/main/java/com/stock/stockmanager/repository/CategorieRepository.java
@@ -3,5 +3,5 @@ package com.stock.stockmanager.repository;
 import com.stock.stockmanager.model.Categorie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategorieRepository extends JpaRepository<Categorie, Long> {
+public interface CategorieRepository extends JpaRepository<Categorie, Integer> {
 }
